requirements.txt
fastapi==0.88.0
sqlalchemy==1.4.45
uvicorn==0.20.0
httpx
psycopg2-binary
python-dotenv
async-sqlalchemy
greenlet